update naming (LB, OFB, OFL) and readme
