+++ examples/LCOE_stack/LCOE_LB_demo.py
@@ -42,7 +42,7 @@ prob = glue.create_setup_OM_problem(
     wind_rose=wind_rose,
 )
 
-if True:  # set true to run one-shot analysis
+if False:  # set true to run one-shot analysis
 
     # setup the latent variables for LandBOSSE and FinanceSE
     cost_wisdem.LandBOSSE_setup_latents(prob, modeling_options)
