+++ test/system/LCOE_stack/test_LCOE_OFB_stack_pyrite.npz
