+++ test/system/LCOE_stack/test_LCOE_OFL_stack_pyrite.npz
